Music, SoundFX, SceneManager, Oh my!
The Screen Manager is almost ready to accomodate the maps we have made, and by "almost ready" I mean I am "almost ready" to transplant the original Map concept into this project. The important thing is that the Screen Manager is talking to all the right entities before I start so there is minimal confusion while I'm reconnecting the guts. I expect to start this tomorrow afternoon.

Part of the process of putting all this together involved putting music and soundFX in the game. I used the Barrel Volcano theme from Super Mario RPG (the best game) as the main soundtrack. I ripped it from a fan site and cleaned it up in Audacity so the song will loop indefinitely. I also added sound effects to the menu and additional title cards for when the game starts. One sound effect is also from SMRPG: the "enter-battle" sound effect; the other is the "cursor-move" sound effect from Final Fantasy 7, but I think this was technically used starting from at least FFVI, and was also used in SMRPG (as Squaresoft, as the company was known then, is responsible for every game).

TODO: Right now the player only has to press "Enter" at any time to jump to the game. Skipping menu animations is generally a good feature, but it does not yet distinguish between "Play" and "Credits". When the Credits scene is made, this will be an easy adjustment. We will just check the position of the cursor along with the pressing of "Enter" as flags for the correct scene transition.