Corrected another Music issue, creating checkpoint
I am about to refactor a big chunk of the screen manager so this is a backup before I do that. This corresponds to the state of things a little more than halfway through Galaga-8 tutorial. After I publish this, I will attempt to move the title cards out of the PlayScreen class and into the Level class.