Map file is being read now but other problems
Map file is being read now. The problem was first that fstream didn't know where the file was located, the full path was required (yuck! why?!) Further, the spaces I have in my own directories (usually I avoid this) caused problems. I had to remove the white space. Now it reads the file. But for some reason, the tiles are all either stacked on top of eachother or the bottom-right of the map is at 16,16 (unlikely). We might need to change how the tileset.png file is turned into a texture