Initial push (WIP), fixed memmory leak
Huge memory leak plugged!
Scenes were being instantiated over eachother when we switched between scenes. This was a terrible oversight partially due to deviating from the tutorial, which had not yet dealt with the problem. A valuable learning experience.

Colliders are about halfway done, will be implemented by end of day. This method will be more abstract than the previous in that it is intended to accomodate rotation.