Colliders are in place and appear correct. Now I set the resolutions
* Colliders are in place and appear correctly.
* Problem was in the constructor, I wasn't setting a member variable.
* This oversight was possible because I use the parameter it gets set with in a chain of IF statements but never actually remembered to set it.
* Now I need to set collision resolution.
* tags (`TriggerType`) are a member variable (enum) of the Collider class.
* I need to set:
      * What happens when a `Player` collides with a `Wall`;
      * What happens when a `Player` collides with a `Web`;
      * What happens when a `Player` collides with `Spikes` or a `Monster`;
      * What happens when a `Bullet` collides witha `Monster`;
      * What happens when a `Bullet` collides with a `Wall`.

I am interested in advice: Does it make sense to set these resolutions in the` Collider` class or in the `PlayScreen` class?

I am thinking of setting up the resolutions in their own functions in the `Collider` class, but calling them only in the `Bullet` and `Player` class, since they are agents of the player.