improvements, bug fixes
* resolving the killing of monsters in the spaghetti-code that is `PlayScreen.cpp` is a difficult task, but I went around refactoring (again) by only allowing collision for monsters whose state is `Active()` and deactivating them when they get hit by a bullet.
Before this was done, deleting a monster on bullet collsion would cause the game to crash because it isn't done accessing it in Render() and/or Update(). Instead of refactoring, I opted not to delete the enemies when they are hit by a bullet, but deactivate them instead. The enemies are automatically deleted when the nex level starts.

* Exits and Rick Tangle's colliders are much smaller now.

*Credits screen is done. It looks great and is great. It sounds great too, it has its own sound effect!

* There is some bug that prevents the high score from being assigned before the player starts a level over once. Also, becasue the start screen is being completely deleted and re-uploaded right now, it is not possible to keep a score after you die. Post to Issues.