HUGE BUG FIXES!
There was a mistake in PlayScreen.cpp for running collision detection between bullets and monsters. The mistake was that the SPACEBAR was required to INITIATE the call, which never should have been. Taking out the if-statement for getting the `space` key completely resolves collision--except for one more thing: Because I am not deleting monsters when they are killed (only disabling them), their invisible colliders are still active, and firing bullets at them makes said bullets dissapear. I resolved this by putting another `if`-statement inside the `ResolveBulletCollsion()` function that checks that the monster is active before doing AABB().

This is a workaround in a short but heavy list of workarounds. A better solution would be to more carefully order the comings-and-goings in PlayScreen's `Update()` function, so we can just delete the monsters when they are hit, instead of waiting for the level to end first.